week 2/week2-agent-tools-function-calling/Day 6/tools - approach 1/requirements.txt
openai>=1.9.0
langchain>=0.2.9
langchain-openai
langchain-community
langchain-chroma
python-dotenv